requirements.txt
fastapi[standard]==0.115.5
pandas==2.2.3
torch==2.2.2
torchvision==0.17.2
numpy==1.26.1
scikit-learn==1.5.2
ultralytics==8.3.49
pillow==11.0.0